backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.6.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
pydantic>=2.5.2
pillow>=10.0.0
python-json-logger>=2.0.7
google-generativeai>=0.3.0